ep_testsuite: Add a maximum wait time to wait_for...
For the various wait_for_stat_{to_be,changed} etc methods, add a
maximum time to wait for, defaulting to 60s. If the stat does not
reach the desired value in that time, abort() the test.

This makes it quicker to detect test failures - instead of waiting
forever (and having to hit some external whole test run timeout).

Change-Id: Iab0d4c3de1a4b7bca8baf279da56f27423c33a18
Reviewed-on: http://review.couchbase.org/44004
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>