MB-13011: Return updated item* on successful arithmetic operation
To facilitate passing additional information about mutations back to
the client, refactor the arithmetic API to return a pointer to the
updated item*. This can then be queried by the frontend (using
get_item_info) to obtain additional information (such as vbucket UUID
and seqno).

Previously CAS was also returned, but this can now be obtained from
get_item_info() so it has been removed from the interface.

Change-Id: If45cc21e04201614b4ff6df75abb762d3623605a
Reviewed-on: http://review.couchbase.org/44714
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dave Rigby <daver@couchbase.com>