MB-13011: Include vBucket UUID & seqno in remove response if requested
Similar to the previous arithmetic change, add a new argument to
remove() in the engine interface which is used to return the mutation
details (vbucket UUID & sequence number) to memcached. These are then
returned back to the client if they have enabled the MUTATION_SEQNO
connection option.

Note this is a breaking change to the engine interface and hence will
require updates to ep-engine.

Change-Id: Ifcfcc0c32cbc190df3c6c26c3bc08461b0c08081
Reviewed-on: http://review.couchbase.org/44757
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>