MB-12887: Unit test for breakpad
Add a unit test for breakpad within memcached.

Adds a new 'engine' which simply crashes when initialized. Testcase checks
that breakpad is correctly when this is loaded, and that the resulting
core file (on Linux) contains useful information. A more limited testcase
on Windows just checks that the minidump is created and is non-zero in
size.

Change-Id: Ie02f2ea8b5ead9162454d41865f6d030937d7070
Reviewed-on: http://review.couchbase.org/45036
Tested-by: buildbot <build@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>