MB-13069: Set range properly when getting checkpoint items
We only update the snapshot range when we see a checkpoint end
message, but replica vbuckets will not contain one of these if
the items are in the open checkpoint. As a result we must update
the range end if we are at the last item in the checkpoint manager.

Change-Id: I4362b1ffb931ccbefbb424a951c08014035e586c
Reviewed-on: http://review.couchbase.org/45234
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>