Return AUTH_STALE to clients for stale connections
Previously the client couldn't detect why it suddenly was
disconnected from the server. This patch make sure we
always send AUTH_STALE when the authentication context is stale
so that the client know that they have to reauthenticate.

Please note that the current version of the server don't support
invalidating the authentication context, so you'll never receive
this response (but you should be prepared to receive it
in the future)

Change-Id: I306dee9aa5a76f8af823dcaeef0ddee6e3f44031
Reviewed-on: http://review.couchbase.org/45265
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>