Add force shutdown mode to logger
In error situations (for example when we have detected SIGABRT /
SIGSEGV) it is desirable to have as much logging inforamation
present. However our logging design buffers log messages in memory,
with a seperate thread performing the I/O to write to disk. We could
attempt to cleanly shutdown the logger in such error situations, but
the state of the logger thread is unknown, or another thread may have
hold the logger mutexes.

To solve this, add a 'force' mode to the logger shutdown() method. If
force==true then any data in the buffers is immediatley flushed to
disk, ignoring any mutexes and the main logger thread. While there is
a risk this may corrupt or fail to print a log message in progress,
this seems a reasonable tradeoff compared to never seeing any of the
buffered log messages.

Change-Id: I126235796710c68a71d3dafc4b317715d452a22c
Reviewed-on: http://review.couchbase.org/45384
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>