MB-13156, MB-12088: Add require_init
Add a new configuration parameter that disables all user
commands while the server (Couchbase Server) is initializing
the node (creating the buckets etc). Until the node is
initialized memcached will only allow the "admin user" to
connect to the cluster and run commands. All other users will
receive a "NOT INITIALIZED" response for all commands except
SASL requests; which will be allowed, but upon a successful
authentication "NOT INITIALIZED" will be returned unless
the SASL authentication was done for the admin user.

Change-Id: Ia25fac29c1e58420ba1da3a7ffa8aa2f5b6b8795
Reviewed-on: http://review.couchbase.org/45553
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>