MB-13173: Only shutdown logger in Breakpad callback if it exists
breakpad_test.py was failing on CentOS 6. This was due to a segfault
in the breakpad callback function (!) caused by dereferencing a null
logger shutdown pointer.

Fix is to check it is non-NULL before tryingt to use it.

(Interestingly this only affects the unit-test which *doesn't* use a
logger extension. In our standard production deployment a logger is
present and so the shutdown method will always be non-NULL).

(cherry picked from commit c7b2064103750f77fe9954b374f4a660232c34ff)

Change-Id: I0ad96ace2f526309bf495903dd32b842615e5eb0
Reviewed-on: http://review.couchbase.org/45767
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>