MB-13202: Get snapshot info from closed chk when open chk is empty
If the open checkpoint doesn't contain any items and we are not in
a partial snapshot then we need to get information from the closed
checkpoint when getting the snapshot info because we want to
resume from the last sequence number in the closed checkpoint.

This situation arises when the dcp consumer recieves a snapshot
marker for a new snapshot and then the connection is closed. In
this case the snapshot start and snapshot end sequence numbers
will reflect the snapshot that was going to be received. When
we reconnect the values will be invalid as a result so we must
get the information from the closed checkpoint. Once the connection
is established the first snapshot marker will overwrite the open
checkpoint because it has 0 items.

Since the checkpoint manager won't always have a closed snapshot
we can safely rely on setting the snapshot start and snapshot end
sequence numeber to the value of the last by sequence number.

Change-Id: I59303ae9f119223a1c9373a57ab234085a9d587a
Reviewed-on: http://review.couchbase.org/46401
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>