Allow testapp to run in parallel on WIN32
Update how testapp launches memcached to allow it to use a automatically
assigned port number instead of hardcoding 11211. This allows multiple
testapp instances (i.e. plain and SSL) to run in parallel, along with all
the other tests.

As a consequence this significantly speeds up the WIN32 "make test"
runtime - halving it on a 2 CPU Windows VM.

Change-Id: I5ae7d732b29f6c08fe72e85487b2770f07dc1831
Reviewed-on: http://review.couchbase.org/46637
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>