MB-13538: Let the audit log be buffered by default
Audit events is typically relatively small and in the case
of a high burst of events doing a flush of the io buffers
for each event would just slow things down. In our model
where we could drop events anyway I think it is an acceptable
risk to use buffered IO for this and only flush after
finishing writing a chunk of events.

Change-Id: Ic4549ba4e0e7a84bd555d6d5dd76cf63407a507a
Reviewed-on: http://review.couchbase.org/47082
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>