MB-13783: mcstat: Correctly consume payload on request failures
If an incorrect key is requested from the server (returning a keylen
of 0), ensure that mcstat correctly consumes any remaining data on the
recv (i.e. an error message). Also print the sent error message to
stderr if the request didn't succeed.

This solves an issue where mcstat would hang indefiniately as it had
incorrectly calculated how much data was left on the socket.

Change-Id: I1df23d5a2c939e184fec5b595f07f703c8c14d08
Reviewed-on: http://review.couchbase.org/47863
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>