MB-13783: Fix potential null-pointer reference
When requesting "default" stats the key is NULL, and calling
fprintf with %s being a NULL pointer. The C standard does NOT
explicitly state what happens if the pointer is a NIL pointer,
BUT it does require that the pointer contains a "NIL" to
terminate the string.

In order to be future-proof we should print out the error code
on errors rather than the payload (which may go away soon)

Change-Id: Ie7b3c5a15b67f31cf7c309147d6c0dba83d69560
Reviewed-on: http://review.couchbase.org/47988
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>