Pass the compaction expiry callback in the context
There was no need for a callback here originally since the code
the callback executd could have just been run after the compaction
code finished. This code doesn't however remove the callback, but
uses it in a way that allows us to not keep a list of expired items.
This should help reduce memory usage when there are a lot of expired
items during a compaction.

Change-Id: Icf27cc0fa231b8834b94c920f36385d0fcf42536
Reviewed-on: http://review.couchbase.org/48337
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>