MB-13745: Ensure audit file is closed when disabled
We can't use the "standard" way of adding events to
the audit log and process them at a later time in the
case when we're disabling the audit, since that will cause
the audit file to be reopened.

"Buffered" audit events needs to be ignored after we've
disabled the audit.

Change-Id: I6efc81375aaa3b78a00b5da4ef5e9c67ddc84bb6
Reviewed-on: http://review.couchbase.org/49495
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>