MB-14374: Properly find the latest rollback point
If we have received a full snapshot on disk then we want to use the
snapshot end sequence number, but if we are in the middle of a
snapshot then we want to use the snapshot start sequence number. We
can figure out what to use by checking the last sequence number
persisted.

Change-Id: I4feeea862628d7c516138e4491aab83b8b05e56a
Reviewed-on: http://review.couchbase.org/49475
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>