Add 'exp_pager_enabled' configuration option
Currently the expiry pager can only be disabled by setting
exp_pager_stime to zero - as this value is special-cased to mean
"never run". However it is sometimes helpful (e.g. in unit tests) to
be able to specify an actual sleep time of zero, so the that the
expiry pager /always/ runs.

Therefore add a new option 'exp_pager_enabled' to explicitly
enable/disable the expiry pager; this then frees up the 'magic' zero
value to actually be used to set the sleep time to zero.

Change-Id: I93dd327371bdb91ef0e3a773a3bc8d2626f425ed
Reviewed-on: http://review.couchbase.org/49807
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>