ep_test_apis: Raise exceptions when attempting to access non-existant or wrong type of stats
The various get_XXX_stat() functions don't return any errors if a user
attempts to access a non-existant stat, or they try to read a string
stat as a int.

Change to use map.at() and stoXXX() which both raise exceptions if the
input is invalid (key doesn't exist / string can't be converted to
int). To facilitate some tests which deliberatly access stats which
may not exist, add new get_XXX_stat_or_default() function which
returns the given default value if the stat doesn't exist.

Change-Id: I38b637f19c11543b60ef75c2f51bae516b6710ec
Reviewed-on: http://review.couchbase.org/49854
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>