Update the engine_testapp API.
The patch adds a new API which delegates the bucket
create/destroy management to the test function.

API-1 is the legacy API in which a bucket is created for the
test function.

API-2 is the new API where each test is handed a engine_test_t
so they can allocate buckets using the test->cfg and new
harness methods.

The driver here is to enable multi-bucket test-cases allowing
the test to create and destroy buckets at their choosing.

Change-Id: I3b884106de6ee7023cafda3e48681cfc2ee19ab3
Reviewed-on: http://review.couchbase.org/50643
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>