Standardise on using engine_loader functions and introduce destroy_engine.
This change primarily introduces a destroy_engine method so that
engines can clean up before unloading. Using engine_loader as the
single API for loading ensures all engine define the method.

ep.so currently does not unload the malloc hooks and we segfault
if memory alloc/dealloc occurs post dlclose of ep.so. Adding the
destroy_engine path allows ep.so to clean up the hooks.

Change-Id: I4ead9f87388110078e495f951ae22cb06c089527
Reviewed-on: http://review.couchbase.org/49978
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>