Make memcached multi-bucket-aware
This is step one.

With this patch all engines will be loaded during startup
so that we know that all symbols will be resolved in the
engine (so that we know that during startup rather than
after running for 3 weeks when someone decides to create
a bucket of type x). If you want to test out an experimental
engine you have to do that through the ewouldblock engine

Change-Id: Idb4ef895dacf2d916b07af4aebca79cba7fda6ba
Reviewed-on: http://review.couchbase.org/50053
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>