Add new error-injection engine: EWOULDBLOCK_engine
The "ewouldblock_engine" allows one to test how memcached responds
when the engine returns EWOULDBLOCK instead of the correct response.

Motivation:

The EWOULDBLOCK response code can be returned from a number of engine
functions, and is used to indicate that the request could not be
immediately fulfilled, and it "would block" if it tried to. The
correct way for memcached to handle this (in general) is to suspend
that request until it is later notified by the engine (via
notify_io_complete()).

However, engines typically return the correct response to requests
immediately, only rarely (and from memcached's POV
non-deterministically) returning EWOULDBLOCK. This makes testing of
the code-paths handling EWOULDBLOCK tricky.

Operation:

This engine, when loaded by memcached proxies requests to a "real"
engine.  Depending on how it is configured, it can simply pass the
request on to the real engine, or artificially return EWOULDBLOCK back
to memcached. The mode can be selected by sending a
`request_ewouldblock_ctl` command (opcode
PROTOCOL_BINARY_CMD_EWOULDBLOCK_CTL). Possible modes are:

* NEXT_N - Make the next_N calls into engine return EWOULDBLOCK. N
           specified by the {value} field.

* RANDOM - Randomly return EWOULDBLOCK. Chance to return EWOULDBLOCK
           is specified as an integer percentage (1,100) in the
           {value} field.

* FIRST  - The first call to a given function will return EWOULDBLOCK,
           with the next (and subsequent) calls to the *same* function
           operating normally. Calling a different function will reset
           back to failing again. In other words, return EWOULDBLOCK
           iif the previous function was not this one.

Change-Id: I64fae62f0b6c27e7b6c13d49fbea83d3261e26aa
Reviewed-on: http://review.couchbase.org/50283
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>