Add cmd_context to connection
Add a new field to the `conn` struct: cmd_context. This is for use as
command-specific data. It's value is reset to NULL between each
command. To allow for cleanup before resetting, a destructor funtion
pointer can be specified which will be called on the cmd_context
before resetting them both.

Change-Id: I53c82119c73f0fffa9a0a7a7c9dfd5ff73646a85
Reviewed-on: http://review.couchbase.org/50383
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>