Convert memcached testapp to C++
Change memcached's testapp to C++ code. Note this is a minimal
conversion to allow the file to compile cleanly as C++, it doesn't
update it to all be ideomatic C++ (for example malloc, calloc et all
are still used for memory management), however in a couple of places
where converting to C++ containers etc was simpler things were
changed.

Change-Id: Ib7ef6cc046edcea9f4b77ab5a8529c5558242265
Reviewed-on: http://review.couchbase.org/50983
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>