Speedup memcached shutdown from 1s to ~instant
Currently to shutdown memcached a global flag (memcached_shutdown) is
set by the various methods to terminate (SIGTERM / SIGINT, shutdown
from engine) and this flag is checked by the clock event handler
(which fires every second). The upshot if this is it typically takes
1s for a request to terminate memcached to actually result in it
stopping. While this is of little consequence for normal production
use; it is an inconvenience for unit testing, as every startup /
shutdown cycle takes a minimum of 1s.

Improve this by breaking the event loop (and terminating memcached) as
soon as the flag is set, i.e. in shutdown_server().

Change-Id: Ifb0763e31b5dd800ed92daf9b04a8e56929d0d5e
Reviewed-on: http://review.couchbase.org/51099
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>