ewouldblock_engine: Add 'Sequence' mode
Add a new fault injection mode to ewouldblock_engine - 'sequence'. In
this mode, the user can specify that the next N engine calls will
return either the normal result or the injected error code, based on
the sequence of bits in the value.

For example, specifying EWBEngineMode_SEQUENCE with a value of 0b0100
means:

- The next two engine calls will return their normal value
  (from the real engine)
- The 3rd call will return the selected error code,
- The 4th will return the normal value from the real engine).

Change-Id: I7d8ac54dd4cf5a85c2c514ce41b38b12a9cd4d05
Reviewed-on: http://review.couchbase.org/51466
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>