MB-13653: Subdoc - handle CAS mismatch and retry when CAS not specified
If the client didn't specify a CAS, we still use CAS internally to
check that we are updating the same version of the document as was
fetched.  However in this case we need to auto-retry in the event of a
concurrent update by some other client.

Add support for this auto-retry, along with a memcached_testapp which
exercises it (making use of the new 'sequence' feature of
ewouldblock_engine).

Change-Id: If5aa84720a9ea4febc8a68fe52c0e3abb2964748
Reviewed-on: http://review.couchbase.org/51467
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>