MB-15331: Automatically bind to "default"
New connections should always be bound to the "default" bucket
if it exists. Previously this was done by bucket_engine in the
"on connect" callback, but now that memcached is multibucket
aware we can do it when we're associating the initial bucket.

If there isn't a bucket named 'default' the "no-bucket" is
selected.

The name of the default bucket was configurable in
bucket-engine, but hardcoded from ns_server. That means there
isn't much point of keeping the flexibility (if we want it
back we can move the setting into memcached.json)

Change-Id: I3c2aaeec826639bafc5bb377d016666e9a24c72d
Reviewed-on: http://review.couchbase.org/52099
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>