MB-13642 Use a strong cipher list by default
If someone wants to use SSL we should try to be "secure by
default", and only allow for using strong ciphers.  Users that
may want to use a less secure cipher list should be allowed to do
so by setting the environment variable COUCHBASE_SSL_CIPHER_LIST
Whenever ns_server allows for specifying the SSL cipher list in
the UI, it will be stored in memcached.json and override these
settings.

Change-Id: I306bd6a22b91806b7cd061e1bead01107451689d
Reviewed-on: http://review.couchbase.org/52145
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>