MB-15363: Handling temp_non_existent_items
In FULL_EVICTION scenario, when a GET or DELETE are called on an
item that does not exist, in the current code base, we create a
TEMP item and issue a background fetch. As the item does not exist,
the temp item is marked as a non_existent_key and is left as is
for the expiry pager to remove (which runs once an hour). If there
are many such operations over a different key set, memory usage
will increase because of all the temp items.

To avoid this, this fix ensures a non_existent_key's temp item
is deleted, before returning ENOENT.

Change-Id: Ic219230f8c0f52a3be7521d266b68a04bdf01079
Reviewed-on: http://review.couchbase.org/52193
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: buildbot <build@couchbase.com>