Perf: Skip logging DEBUG event if debug disabled
In update_event() we log (at DEBUG level) what the event has been
changed to. This occurs very frequently (on every command received),
and has a reasonable high cost as we create a SyslogEvent (see
logger_log_wrapper) for every log() call; even if the severity of the
event is ultimately not logged to a file.

Change-Id: I1e048666a6f8454902b2173cbe9a8abedc1f9d00
Reviewed-on: http://review.couchbase.org/52368
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: buildbot <build@couchbase.com>