testapp: Check for exit status of memcached
When terminating memcached at the end of the test run, verify the exit
status is a clean shutdown. Along with being a good test in it's own
right, this allows analysis tools which interpose a non-zero exit code
if they detected errors (e.g. Valgrind, ThreadSanitizer) to mark a
test as failed.

Change-Id: I93c7b667924d4cf25e29de03acd02a2cf22de376
Reviewed-on: http://review.couchbase.org/52483
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>