Remove CQ_ITEM freelist from threads.cc
threads.cc implements a freelist for freed CQ_ITEM objects. This has a
number of issues, however:

1) Access to the freelist is serialized behind a single mutex;
affecting both allocation and free'ing of CQ_ITEMS.

2) The freelist isn't correctly free'd back to the OS on process exit,
making it hard to verify it is leak-free.

The second of these is the most problematic, and given that we now use
a memory allocator (jemalloc / TCMalloc) which supports thread-local
allocations which do /not/ require acquiring a global lock in the
common case there is questionable value in this freelist now.

Removing the freelist and just using malloc() / free() directly.

Change-Id: Ie041641e63de9d6010fc28e6d7fe1b8d849ecc4c
Reviewed-on: http://review.couchbase.org/52622
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dave Rigby <daver@couchbase.com>