Free settings struct on shutdown
Free all of the settings struct (and any sub-structures when shutting
down memcached. The missing free() of settings->admin is a 'real'
memory leak which would be triggered every time the admin username was
changed; however it essentially never changes after startup in a real
system and is only a leak of <10 bytes, so not a critical issue (and
hence no need for backporting).

Change-Id: I843338a84de83a196688f11abae1ae776fd83008
Reviewed-on: http://review.couchbase.org/52626
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dave Rigby <daver@couchbase.com>