Subdoc API: Always free dynamic_buffer at end of command
The dynamic_buffer member of the connection object was historically
only used for the complete response from some commands. As such it was
always freed via the write_and_free() method, which (as the name
suggests) writes the given buffer to the client and frees the buffer
afterwards. Beginning with the subdocument API, dynamic_buffer is also
used as temporary storage for decompressing a compressed JSON input
document. As this buffer is never directly written back to the client
then write_and_free() was not used and hence the buffer could leak.

Solve this by always free()ing dynamic_buffer at the end of a command,
or when the connection is closed.

Change-Id: I1fc4c651c699718c1ce47168f8c709f372982918
Reviewed-on: http://review.couchbase.org/52657
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dave Rigby <daver@couchbase.com>