MB-15166: Takeover to throttle front end load if necessary
If a takeover stream is taking too long in the "takeover-send"
phase because of its checkpoint queues being heavily backed up
due to a very high front end, then it will now throttle the
front end by telling the client to back-off.

This could cause tmpOOMs during rebalance in certain scenarios
but would reduce the risk of rebalance remaining stuck in the
midst of very high front end ops (mutations).

Change-Id: I90acc37fb3fd6a95b8c1b768a47421c45490b447
Reviewed-on: http://review.couchbase.org/51889
Reviewed-by: Chiyoung Seo <chiyoung@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>