basic_engine_testsuite: Relax get_item_info expiry time assert
The get_item_info test in basic_engine_testsuite tries to check that
the returned expiry time is the same as the requested value, however
this is an incorrect use of the API.

The input value to item_allocate() is a memcache protocol time
(i.e. <30 days; treat as relative time from now, otherwise treat as
time since Unix epoch), but the `exptime` member of the item_info
struct is a rel_time_t - i.e. a time relative to when the memcached
server starts. For example if item_allocate() is called with an expiry
of '1', but the server has already been up for 5 seconds then
get_item_info() will return a exptime of '6' (assuming no time has
passed between the two calls).

Therefore relax the assert to just check that the returned value is
greater or equal to the requested value.

Fixes intermittant test failure in basic_engine_testsuite:

    assertion failed [ii.exptime == exp] at memcached/testsuite/basic/basic_engine_testsuite.cc:451
    Called from:
        platform/libplatform.so.0.1.0(cb_assert_die+0x66) [0x504bf66]
        basic_engine_testsuite.so() [0x7405c14]
        memcached/engine_testapp() [0x403570]
        /lib/x86_64-linux-gnu/libc.so.6(__libc_start_main+0xed) [0x56eb76d]
        memcached/engine_testapp() [0x4038fd]
    Killed

Change-Id: I18f765ae09772fe9c1dad631865d71c50f3cb739
Reviewed-on: http://review.couchbase.org/52706
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>