engine_testapp: Replace system() with execvp/CreateProcess
Change engine_testapp to spawn the individual test instances using
execvp (and the WIN32 equivilent) instead of system(). The main
motivation for this is that the shell (/bin/bash) used by sytem() has
a number of memory leaks, which create false positives when running
under Valgrind.

This change also speeds up test running slightly, as there is no
longer an additional intermediate shell process required to invoke the
tests.

Change-Id: I15cbd0923b917a547160bdbb1df28d80a469a645
Reviewed-on: http://review.couchbase.org/52856
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>