Remove use of std::thread due to buggy MSVC 2013
Under MSVC 2013 std::thread when used for the default_engine scrubber
task causes a deadlock - the creation of the std::thread never completes
(we never see ScrubberTask::run() called), and the parent thread is
blocked waiting for the thread to be created.

Replacing with cb_thread (which uses the CreateThread() primitive on
WIN32) solves the deadlock.

Change-Id: Ie0949cb61b28a0d23da5b52d348139f931c74a22
Reviewed-on: http://review.couchbase.org/52871
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>