MB-13642: Don't use env variable to sett cipher list (backport)
It turns out that it is hard to get the environment variable
populated from the startup script all the way down to memcached
on Microsoft Windows. Instead of making a "one off" for windows
by using the registry, lets just configure this through a
file located in $ROOT/etc/security.

Change-Id: I8eaec8369bcfcecc91d8457d03e7e1240927a12a
Reviewed-on: http://review.couchbase.org/52970
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: buildbot <build@couchbase.com>