testapp: Fix race condition in opening memcached port file
On WIN32 (maybe elsewhere) testapp occasionally fails to open the
memcached port file; which causes all subsequent unit tests to fail. The
issue seems to be related to the seperate access() and fopen() calls made
to detect when the file is created. Occasionally the fopen() will fail
with EPERM, even though the access() call must have succeeded.

The exact issue here is unknown, I assume on WIN32 it is possible that the
file path or similar exists before the actual file is accessible.
Regardless, simply remove the call to access() and just (repeatedly) try
to fopen() it, continuing when the call succeeds. This appears to solve
the intemittant failure on WIN32.

Change-Id: I0cb62167938843e443324fb460e6ef31cfa180b1
Reviewed-on: http://review.couchbase.org/53026
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>