MB-15573: Remove unpredicatable code from warmup and improve DGM scenarios.
Force 1 active vbucket to the top of the shard's list of VBs to ensure that
bad DGM setups on nodes with a low number of vbuckets doesn't get trashed
by replica vbuckets.

Use C++11 features so that we generate a repeatable psudeo random sequence
and use the distribution API to weight this between active/replica buckets.

A subsequent patch can be made to introduce config values for the weights.

Change-Id: Ie729c5f5e7cbd16e85bba2065aa9c768ffd12774
Reviewed-on: http://review.couchbase.org/53030
Tested-by: buildbot <build@couchbase.com>
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>