ConnManager: Fix connection leak on engine shutdown
When ep_engine is shutdown, there is the potential to leak connections
(aka cookies) from the engine. This can occur if the connection has
has it's refcount incremented (for example if the connection has
performed a DCP_OPEN), as the refcount decrement is done lazily by the
ConnManager class.

Solve this problem by setting the completeBeforeShutdown constructor
argument on GlobalTask to true, ensuring that we complete any
outstanding connection cleanup upon shutdown.

This fixes a memory leak in a number of DCP unit tests, including (but
not limited to) test_dcp_consumer_flow_control_buf_sz:

    1,920 bytes in 10 blocks are definitely lost in loss record 22 of 28
       at 0x524CA6C: je_calloc (jemalloc.c:1527)
       by 0x40656E: calloc (alloc_hooks.c:82)
       by 0x405D9F: create_mock_cookie (mock_server.c:420)
       by 0x740BC31: test_dcp_consumer_flow_control_buf_sz(engine_interface*, engine_interface_v1*) (ep_testsuite.cc:3665)
       by 0x403393: execute_test (engine_testapp.cc:1084)
       by 0x403393: main (engine_testapp.cc:1426)

Change-Id: I3153205fd446bc82b357c5bd412907697f8329b5
Reviewed-on: http://review.couchbase.org/53128
Reviewed-by: abhinav dangeti <abhinav@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung@couchbase.com>